# Conflicts:
#	api-spanboon/src/api/controllers/GuestController.ts
#	api-spanboon/src/api/controllers/PageController.ts

+++ api-spanboon/src/api/services/DeleteUserSerivce.ts
@@ -24,7 +24,6 @@ import { SocialPostLogsService } from './SocialPostLogsService';
 import { SocialPostService } from './SocialPostService';
 import { SearchHistoryService } from './SearchHistoryService';
 import { PostsService } from './PostsService';
-import { PostsGalleryService } from './PostsGalleryService';
 import { PostsCommentService } from './PostsCommentService';
 import { PageUsageHistoryService } from './PageUsageHistoryService';
 import { PageSocialAccountService } from './PageSocialAccountService';
@@ -60,7 +59,6 @@ export class DeleteUserService {
         private fulfillmentCaseService: FulfillmentCaseService,
         private pageObjectiveService: PageObjectiveService,
         private pageConfigService: PageConfigService,
-        private postsGalleryService: PostsGalleryService,
         private pageAccessLevelService: PageAccessLevelService
 
     ) {
@@ -68,10 +66,6 @@ export class DeleteUserService {
 
     public deleteUser(userObjId: string): Promise<any> {
         return new Promise(async (resolve, reject) => {
-            const findAssetOwn = await this.assetService.findOne({ userId: userObjId });
-            if (findAssetOwn !== undefined) {
-                await this.assetService.deleteMany({ userId: userObjId });
-            }
             const findAuthenOwn = await this.authenticationIdService.findOne({ user: userObjId });
             if (findAuthenOwn !== undefined) {
                 await this.authenticationIdService.deleteMany({ user: userObjId });
@@ -118,61 +112,77 @@ export class DeleteUserService {
             if (socialPostLogsOwn !== undefined) {
                 await this.socialPostLogsService.deleteMany({ user: userObjId });
             }
-
-            const postOwn = await this.postsService.findOne({ ownerUser: userObjId });
-            if (postOwn !== undefined) {
-                await this.postsService.deleteMany({ ownerUser: userObjId });
-            }
-            const postCommentOwn = await this.postsCommentService.findOne({ user: userObjId });
-            if (postCommentOwn !== undefined) {
-                await this.postsCommentService.deleteMany({ user: userObjId });
-            }
-            const pageUsageOwn = await this.pageUsageHistoryService.findOne({ userId: userObjId });
-            if (pageUsageOwn !== undefined) {
-                await this.pageUsageHistoryService.deleteMany({ userId: userObjId });
-            }
             const findAccessLevel1St = await this.pageAccessLevelService.findOne({ user: ObjectID(userObjId) });
-            // delete
-
-            if (findAccessLevel1St !== undefined && findAccessLevel1St.level === 'OWNER') {
-                const findPageOwn_1 = await this.pageService.findOne({ ownerUser: ObjectID(findAccessLevel1St.user) });
-                const postOwn_1 = await this.postsService.findOne({ pageId: findAccessLevel1St.page });
-                if (postOwn_1 !== undefined) {
-                    await this.postsGalleryService.deleteMany({ pageId: findAccessLevel1St.page });
-                }
+            let findOwnerPage = undefined;
+            if(findAccessLevel1St !== undefined){
+                findOwnerPage = await this.pageService.findOne({_id:findAccessLevel1St.page,ownerUser: ObjectID(userObjId) });
+            }
+            if(findOwnerPage !== undefined){
+                const findPageOwn_1 = await this.pageService.findOne({ _id: ObjectID(findOwnerPage.id) });
                 const searchHistoryOwn = await this.searchHistoryService.findOne({ userId: userObjId });
                 if (searchHistoryOwn !== undefined) {
-                    await this.searchHistoryService.deleteMany({ resultId: findAccessLevel1St.page });
+                    await this.searchHistoryService.deleteMany({ resultId: findPageOwn_1.id});
                 }
-                const pageConfigOwn = await this.pageConfigService.findOne({ page: findAccessLevel1St.page });
+                const pageConfigOwn = await this.pageConfigService.findOne({ page:findPageOwn_1.id });
                 if (pageConfigOwn !== undefined) {
-                    await this.pageConfigService.deleteMany({ page: findAccessLevel1St.page });
+                    await this.pageConfigService.deleteMany({ page: findPageOwn_1.id });
                 }
-                const pageObjectiveOwn = await this.pageObjectiveService.findOne({ pageId: findAccessLevel1St.page });
+                const pageObjectiveOwn = await this.pageObjectiveService.findOne({ pageId: findPageOwn_1.id});
                 if (pageObjectiveOwn !== undefined) {
-                    await this.pageObjectiveService.deleteMany({ pageId: findAccessLevel1St.page });
+                    await this.pageObjectiveService.deleteMany({ pageId: findPageOwn_1.id});
+                    await this.assetService.deleteMany({ userId: userObjId });
                 }
-                const pageFulfillmentOwn = await this.fulfillmentCaseService.findOne({ pageId: findAccessLevel1St.page });
+                const pageFulfillmentOwn = await this.fulfillmentCaseService.findOne({ pageId: findPageOwn_1.id });
                 if (pageFulfillmentOwn !== undefined) {
-                    await this.fulfillmentCaseService.deleteMany({ pageId: findAccessLevel1St.page });
+                    await this.fulfillmentCaseService.deleteMany({ pageId: findPageOwn_1.id});
                 }
-                const pageSocialAccOwn = await this.pageSocialAccountService.findOne({ page: findAccessLevel1St.page });
+                const pageSocialAccOwn = await this.pageSocialAccountService.findOne({ page: findPageOwn_1.id });
                 if (pageSocialAccOwn !== undefined) {
                     await this.pageSocialAccountService.deleteMany({ ownerPage: userObjId });
                 }
-                const socialPostOwn = await this.socialPostService.findOne({ pageId: findAccessLevel1St.page });
+                const socialPostOwn = await this.socialPostService.findOne({ pageId: findPageOwn_1.id});
                 if (socialPostOwn !== undefined) {
-                    await this.socialPostService.deleteMany({ pageId: findAccessLevel1St.page });
+                    await this.socialPostService.deleteMany({ pageId: findPageOwn_1.id});
                 }
                 if (findPageOwn_1 !== undefined) {
-                    await this.pageAccessLevelService.deleteMany({ page: findAccessLevel1St.page });
+                    await this.pageAccessLevelService.deleteMany({ page: findPageOwn_1.id });
                 }
                 if (findPageOwn_1 !== undefined) {
-                    await this.pageService.deleteMany({ _id: findAccessLevel1St.page });
+                    await this.pageService.deleteMany({ _id: findPageOwn_1.id });
+                }
+            }else if(findOwnerPage === undefined){
+                const postOwn = await this.postsService.findOne({pageId:null, ownerUser: userObjId });
+                if (postOwn !== undefined) {
+                    await this.postsService.deleteMany({ ownerUser: userObjId });
+                }
+                const postCommentOwn = await this.postsCommentService.findOne({ user: userObjId });
+                if (postCommentOwn !== undefined) {
+                    await this.postsCommentService.deleteMany({ user: userObjId });
+                }
+            }
+            let findOwnerLevel1St = undefined;
+            if(findAccessLevel1St !== undefined){
+                findOwnerLevel1St = await this.pageAccessLevelService.findOne({page:findAccessLevel1St.page,level:findAccessLevel1St.level});
+            }
+            // delete
+
+            if (findAccessLevel1St !== undefined && findAccessLevel1St.level === 'OWNER') {
+                await this.pageAccessLevelService.deleteMany({ user: ObjectID(userObjId) });
+                const pageUsageOwn = await this.pageUsageHistoryService.findOne({ userId: userObjId });
+                if (pageUsageOwn !== undefined) {
+                    await this.pageUsageHistoryService.deleteMany({ userId: userObjId });
                 }
             } else if (findAccessLevel1St !== undefined && findAccessLevel1St.level !== 'OWNER') {
-                await this.pageAccessLevelService.delete({ page: findAccessLevel1St.page, level: findAccessLevel1St.level, user: ObjectID(userObjId) });
+                const ownerPage = await this.pageService.findOne({_id:findOwnerLevel1St.page});
+                const query = {pageId:ObjectID(findAccessLevel1St.page)};
+                const newValues = {$set:{ownerUser:ObjectID(ownerPage.ownerUser)}};
+                const updatePermission = await this.postsService.updateMany(query,newValues);
+                if(updatePermission){
+                    await this.pageAccessLevelService.delete({ page: findAccessLevel1St.page, level: findAccessLevel1St.level, user: ObjectID(userObjId) });
+                }
+
             }
+
             // deleteOne User
             const userOwn = await this.userService.findOne({ _id: userObjId });
             if (userOwn) {
